Jmanko/onem 20073 (#5)
* ONEM-20073: Fix compilation warnings, add -Werror flag

Most of warnings have been fixed, some were disabled:
-Wcast-align - disabled
-Wformat=2 - changed to 1 (2 enables -Wformat-nonliteral, nonliteral format strings are used in Angle)
-Wpragmas - disabled (macros for disabling case fallthrough warnings fail in DerivedSources)

* Remove .orig files - leftover from applying patches

+++ Source/WebCore/platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.cpp
@@ -1061,7 +1061,7 @@ bool MediaPlayerPrivateGStreamerMSE::supportsCodec(String codec)
         if (codecMatchesPattern) {
             if (codec.startsWith("vp09")) {
                 auto fields = codec.split(".");
-                for (int i = 1; codecMatchesPattern && i < fields.size(); ++i) {
+                for (size_t i = 1; codecMatchesPattern && i < fields.size(); ++i) {
                     bool ok;
                     int val = fields[i].toInt(&ok);
                     codecMatchesPattern = ok && val < 99;
